refactor: move types to core/types
