+++ tests/validationTri.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest'
 import { canPlace, isTriInBounds } from '../src/utils/validation'
-import type { PlacedPiece, PiecesConfig } from '../src/types'
+import type { PlacedPiece, PiecesConfig } from '../src/core/types'
 import { toTriKey, toTriEdgeKey } from '../src/core/utils/coordinateKey'
 import piecesConfig from '../src/data/pieces-config.json'
